fix: componentizando pagina controle de cadastro
